pyproject.toml
[tool.mypy]
python_version = "3.10"
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
ignore_missing_imports = true
disallow_untyped_defs = true
strict_optional = true
strict = true

[[tool.mypy.overrides]]
module = "c2cgeoportal_admin.*"
disallow_untyped_defs = false

[tool.black]
line-length = 110
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 110
known_local_folder = ["c2cgeoportal_commons", "c2cgeoportal_geoportal", "c2cgeoportal_admin", "geomapfish_qgisserver", "{{cookiecutter.package}}_geoportal"]

[tool.poetry]
name = "c2cgeoportal"
version = "0.0.0"
description = "Not used"
authors = []

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
alembic = "1.13.3" # geoportal
"c2c.template" = "2.4.2" # geoportal
c2cgeoform = "2.4.0" # commons
colander = "2.0" # commons, admin
ColanderAlchemy = "0.3.4" # commons
cookiecutter = "2.6.0" # geoportal
deform = "2.0.15" # commons, admin
defusedxml = "0.7.1" # geoportal
"dogpile.cache" = "1.3.3" # geoportal
Fiona = "1.10.1" # geoportal raster
GeoAlchemy2 = "0.15.2" # commons, geoportal
geojson = "3.1.0" # geoportal
getitfixed = "1.0.29" # geoportal
isodate = "0.6.1" # geoportal
Mako = "1.3.5" # geoportal
OWSLib = "0.31.0" # geoportal
papyrus = "2.6.2" # commons, geoportal
passwordgenerator = "1.5.1" # geoportal
psycopg2 = "2.9.9" # geoportal
pycryptodome = "3.21.0" # geoportal
pyproj = "3.7.0" # admin, other?
pyotp = "2.9.0" # geoportal
pyramid = "2.0.2" # geoportal
pyramid_debugtoolbar = "4.12.1" # geoportal
pyramid-jinja2 = "2.10.1" # admin
pyramid_mako = "1.1.0" # geoportal
pyramid_multiauth = "1.0.2" # geoportal
pyramid_tm = "2.5" # geoportal
python-dateutil = "2.9.0.post0" # geoportal
PyYAML = "6.0.1" # geoportal
rasterio = "1.4.1" # geoportal raster
requests = "2.32.3" # geoportal
redis = "5.1.0" # geoportal cache
Shapely = "2.0.6" # geoportal
SQLAlchemy = "2.0.35"
transaction = "5.0" # commons, geoportal
translationstring = "1.4" # admin
"zope.event" = "5.0" # commons, geoportal
c2cwsgiutils = { version = "6.0.8", extras = ["broadcast", "standard", "oauth2", "debug"] }
oauthlib = "3.2.2"
tilecloud = "1.12.3" # geoportal
azure-storage-blob = "12.23.1"
# simple_openid_connect = '1.0.1' # geoportal
simple_openid_connect = { git = "https://github.com/sbrunner/py_simple_openid_connect.git", branch = "allows-pkce" } # geoportal
pkce = '1.0.3' # geoportal
basicauth = "1.0.0"
prospector = { extras = ["with_mypy", "with_bandit", "with_pyroma"], version = "1.12.0" }
prospector-profile-duplicated = "1.5.0"
prospector-profile-utils  = "1.9.0"

[tool.poetry.group.dev.dependencies]
Babel = "2.16.0" # i18n
beautifulsoup4 = "4.12.3" # admin tests
Jinja2 = "3.1.4" # c2c.template
lingva = "5.0.4" # i18n
PasteScript = "3.6.0" # geoportal pcreate
polib = "1.2.0" # i18n
pykwalify = "1.8.0" # Validate config
pytest = "8.3.3" # tests
WebTest = "3.0.1" # admin tests
types-pytz = "2024.2.0.20241003"
types-pyyaml = "6.0.12.20240917"
types-python-dateutil = "2.9.0.20241003"
types-requests = "2.32.0.20240914"
types-setuptools = "75.1.0.20240917"
mappyfile = "1.0.2"
# To be able to upgrade project from version <= 2.8
c2cciutils = { version = "1.5.8", extras = ["checks"] }
pre-commit = "3.8.0"
mapbox-vector-tile = "2.1.0"
responses = "0.25.3"
waitress = "3.0.0"
